feat(react-client): export `processResponse` and `resolveResponse` under `unstable__` prefix
